+++ src/android/Operation.java
@@ -8,7 +8,6 @@ public class Operation {
   public String type;
   public JSONArray args;
   public CallbackContext callbackContext;
-  public BluetoothDevice device;
 
   public Operation(String type, JSONArray args, CallbackContext callbackContext) {
     this.type = type;