+++ src/ios/BluetoothLePlugin.h
@@ -15,6 +15,14 @@
   int requestId;
   NSMutableDictionary* requestsHash;
   NSMutableDictionary* servicesHash;
+
+  BOOL writeQIsRunning;
+  int writeQtype;
+  NSInteger writeQLocation;
+  NSInteger writeQLength;
+  NSInteger writeQChunkSize;
+  NSData *writeQData;
+  CBCharacteristic *currentWriteCharacteristic;
 }
 
 - (void)initialize:(CDVInvokedUrlCommand *)command;
@@ -54,6 +62,7 @@
 - (void)requestPermission:(CDVInvokedUrlCommand *)command;
 - (void)isLocationEnabled:(CDVInvokedUrlCommand *)command;
 - (void)requestLocation:(CDVInvokedUrlCommand *)command;
+- (void)retrievePeripheralsByAddress:(CDVInvokedUrlCommand *)command;
 
 - (void)initializePeripheral:(CDVInvokedUrlCommand *)command;
 - (void)addService:(CDVInvokedUrlCommand *)command;